Update app.js
To fix problems arising due to the asynchronous nature of .listen(). Really it should be doing that console.log call inside a callback passed to listen. When the host is added, it tries to do a DNS lookup, which is async. So when that line tries to fetch the address, there isn't one yet because the DNS request is running, so it crashes.